separate fake pod controller processor fromt the test
