+++ pkg/kube/pod_controller_test.go
@@ -24,7 +24,6 @@ import (
 	. "gopkg.in/check.v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/fake"
 )
 
@@ -41,52 +40,52 @@ func (s *PodControllerTestSuite) SetUpSuite(c *C) {
 	os.Setenv("POD_NAMESPACE", podControllerNS)
 }
 
-// FakePodControllerProcessor is a fake pod controller processor
-type FakePodControllerProcessor struct {
-	inWaitForPodReadyNamespace string
-	inWaitForPodReadyPodName   string
-	waitForPodReadyErr         error
+// // FakePodControllerProcessor is a fake pod controller processor
+// type FakePodControllerProcessor struct {
+// 	inWaitForPodReadyNamespace string
+// 	inWaitForPodReadyPodName   string
+// 	waitForPodReadyErr         error
 
-	inWaitForPodCompletionNamespace string
-	inWaitForPodCompletionPodName   string
-	waitForPodCompletionErr         error
+// 	inWaitForPodCompletionNamespace string
+// 	inWaitForPodCompletionPodName   string
+// 	waitForPodCompletionErr         error
 
-	inDeletePodNamespace string
-	inDeletePodPodName   string
-	inDeletePodOptions   metav1.DeleteOptions
-	deletePodErr         error
+// 	inDeletePodNamespace string
+// 	inDeletePodPodName   string
+// 	inDeletePodOptions   metav1.DeleteOptions
+// 	deletePodErr         error
 
-	inCreatePodCli     kubernetes.Interface
-	inCreatePodOptions *PodOptions
-	createPodRet       *corev1.Pod
-	createPodErr       error
-}
+// 	inCreatePodCli     kubernetes.Interface
+// 	inCreatePodOptions *PodOptions
+// 	createPodRet       *corev1.Pod
+// 	createPodErr       error
+// }
 
-func (f *FakePodControllerProcessor) createPod(_ context.Context, cli kubernetes.Interface, options *PodOptions) (*corev1.Pod, error) {
-	f.inCreatePodCli = cli
-	f.inCreatePodOptions = options
-	return f.createPodRet, f.createPodErr
-}
+// func (f *FakePodControllerProcessor) createPod(_ context.Context, cli kubernetes.Interface, options *PodOptions) (*corev1.Pod, error) {
+// 	f.inCreatePodCli = cli
+// 	f.inCreatePodOptions = options
+// 	return f.createPodRet, f.createPodErr
+// }
 
-func (f *FakePodControllerProcessor) waitForPodCompletion(ctx context.Context, namespace, podName string) error {
-	f.inWaitForPodCompletionNamespace = namespace
-	f.inWaitForPodCompletionPodName = podName
-	return f.waitForPodCompletionErr
-}
+// func (f *FakePodControllerProcessor) waitForPodCompletion(ctx context.Context, namespace, podName string) error {
+// 	f.inWaitForPodCompletionNamespace = namespace
+// 	f.inWaitForPodCompletionPodName = podName
+// 	return f.waitForPodCompletionErr
+// }
 
-func (f *FakePodControllerProcessor) waitForPodReady(ctx context.Context, namespace, podName string) error {
-	f.inWaitForPodReadyPodName = podName
-	f.inWaitForPodReadyNamespace = namespace
-	return f.waitForPodReadyErr
-}
+// func (f *FakePodControllerProcessor) waitForPodReady(ctx context.Context, namespace, podName string) error {
+// 	f.inWaitForPodReadyPodName = podName
+// 	f.inWaitForPodReadyNamespace = namespace
+// 	return f.waitForPodReadyErr
+// }
 
-func (f *FakePodControllerProcessor) deletePod(_ context.Context, namespace string, podName string, opts metav1.DeleteOptions) error {
-	f.inDeletePodNamespace = namespace
-	f.inDeletePodPodName = podName
-	f.inDeletePodOptions = opts
+// func (f *FakePodControllerProcessor) deletePod(_ context.Context, namespace string, podName string, opts metav1.DeleteOptions) error {
+// 	f.inDeletePodNamespace = namespace
+// 	f.inDeletePodPodName = podName
+// 	f.inDeletePodOptions = opts
 
-	return f.deletePodErr
-}
+// 	return f.deletePodErr
+// }
 
 func (s *PodControllerTestSuite) TestPodControllerStartPod(c *C) {
 	ctx := context.Background()
